refactor: hoist pure read-model helpers into sqlx_repo (phase 1)
Move the 17 byte-identical, sqlx-free read-model/validation functions
(plus the shared `IncludeSpec`) out of postgres_repo and sqlite_repo into
a new `sqlx_repo::read_model` submodule and import them from both backends.

These functions (schema resolution, write-plan validation, row-version
arithmetic, key/patch reconciliation, identifier quoting, etc.) carried no
database types, so each fix previously had to land twice and could drift.

Net ~283 lines removed (two ~316-line copies deleted, one ~336-line shared
copy added). No public API change; no behavior change — every function was
verified byte-identical before the move.
